+++ product-service/src/main/java/com/keiko/productservice/service/product/ProductRatingService.java
@@ -5,9 +5,9 @@ import com.keiko.productservice.entity.Product;
 import java.util.List;
 
 public interface ProductRatingService {
-    List<Product> findProductsRatingLessThan (Double averageAssessment, Boolean sortByAscend);
+    List<Product> findProductsRatingLessThan (Float averageAssessment, Boolean sortByAscend);
 
-    List<Product> findProductsRatingMoreThan (Double averageAssessment, Boolean sortByAscend);
+    List<Product> findProductsRatingMoreThan (Float averageAssessment, Boolean sortByAscend);
 
     List<Product> findProductsRatingRange (Float minAverageAssessment, Float maxAverageAssessment, Boolean sortByAscend);
 }
